add docker files

+++ bin/generate-env-file.js
@@ -1,11 +1,13 @@
 #!/usr/bin/env node
 
-const wpms = require('./index');
+const wpms = require('wpms2');
 const fs = require('fs');
 
-wpms.boot().then(() => {
-  fs.writeFile('app/configs/env.json', JSON.stringify(wpms.configs, null, 2), (err) => {
-    if (err) throw err;
-    console.log('done!'); // eslint-disable-line no-console
+if (!fs.existsSync('/app/configs/env.json')) {
+  wpms.boot().then(() => {
+    fs.writeFile('app/configs/env.json', JSON.stringify(wpms.configs, null, 2), (err) => {
+      if (err) throw err;
+      console.log('env.json created'); // eslint-disable-line no-console
+    });
   });
-});
+}
